Fix an issue causing the spawn of an additional child process.
Consider the following scenario:
file content:
a\n
b\n
c\n

If we start counting from 1 and increment each time a newline is found
we get 4.

If the file is *not* properly formatted, there *could* be an incompleate
command description between the last newline and the last symbol of the file.
Luckily, we work with well formated data.