tova zashto raboti bukvalno ne razbiram?
```A pipe whose read or write end has been closed is considered widowed.  Writing on such a pipe causes the writing process to receive a SIGPIPE signal.  Widowing a pipe is the only way to deliver end-of-file to a reader: after the reader consumes any buffered data, reading a widowed pipe returns a zero count.```